requirements.txt
PyGithub==1.59.1
python-dotenv==1.0.0
langchain==0.1.0
langchain-ollama==0.1.0
